Update README.md
In this report, I am analizing data on admissions to hospital and deaths in Scotland from unintentional injuries and assaults covering a ten year period (2011 - 2020). The analysis is going to be centered on the task of exploring which types of injures were the most common cause for admissions according to age and sex and on finding the death rates in admissions resulted for those main causes.